requirements.txt
Django==1.7
South==1.0
basicauth==0.2
dj-database-url==0.2.1
dj-static==0.0.6
django-extensions==1.4.0
django-social-auth==0.7.28
djangorestframework==2.4.4
djorm-ext-core==0.7
djorm-ext-expressions==0.6
djorm-ext-pgarray==0.10
enum34==1.0
evernote==1.25.0
gunicorn==19.1.1
httplib2==0.9
oauth2==1.5.211
psycopg2==2.5.4
python-oauth2==0.7.0
python-openid==2.2.5
requests==2.4.3
six==1.8.0
static==0.4
static3==0.5.1
thrift==0.9.1
wsgiref==0.1.2
